Import alternate bug tracker info from MetaCPAN into RT [rt.cpan.org #80020]
This patch adds the queue Attribute 'DistributionBugtracker' which is
populated using data from the MetaCPAN API.

A two parse approach is taken, one to set or update the information from
MetaCPAN into the RT database and a second that removes the information
from queues that no longer require it.

This work has been sponsored by Shadowcat Systems.

[TRS: Relies on complimentary patches to RT::BugTracker and
RT::BugTracker::Public.  I replaced a call to the 3.8-era LimitAttribute
with the explicit Join and Limits.]

Signed-off-by: Thomas Sibley <trs@bestpractical.com>