Filter out dists with bugtrackers pointing to rt.cpan.org on the MetaCPAN API side
This greatly reduces the number of results we have to fetch over the
wire.  Note that releases with a non-rt.cpan.org mailto address and an
rt.cpan.org web address are intentionally still be returned.

In testing, 5min is more than enough time to process the scrolled set
even with the added query complexity.