Improve query path for finding ticket custom fields
Significantly reduces the overhead imposed on PostgreSQL by improving
the query path used to locate the queue's ticket custom fields 'Broken
in' and 'Fixed in'.  The query(s) for this were the most expensive parts
of the entire process.

Along with this change there should be an index created on
ObjectCustomFields.objectid and RT should be updated/patched to include
dcd729513 - Merge Branch '5.0/use-bind-values' into 5.0-trunk